[ISSUE #12] - fixed slug undefined and building fail
